API token card appears in User or Settings depending on auth type
